Darwin

+++ ApiFirmaElectronica/src/main/java/com/ec/alpha/requestfirma/Validity.java
@@ -1,20 +1,21 @@
 package com.ec.alpha.requestfirma;
 
 public class Validity {
-	  public int not_before;
-	    public int not_after;
-		public int getNot_before() {
+	  public long not_before;
+	    public long not_after;
+		public long getNot_before() {
 			return not_before;
 		}
-		public void setNot_before(int not_before) {
+		public void setNot_before(long not_before) {
 			this.not_before = not_before;
 		}
-		public int getNot_after() {
+		public long getNot_after() {
 			return not_after;
 		}
-		public void setNot_after(int not_after) {
+		public void setNot_after(long not_after) {
 			this.not_after = not_after;
 		}
+	
 	    
 	    
 }
